feat(nuxi): expose `runCommand`
