+++ packages/nuxi/bin/nuxi.mjs
@@ -1,3 +1,3 @@
 #!/usr/bin/env node
 process._startTime = Date.now()
-import('../dist/index.mjs').then(r => (r.default || r).main())
+import('../dist/cli.mjs').then(r => (r.default || r).main())
